pyproject.toml
[project]
name = "ksp-naboj"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-cleanup>=9.0.0",
    "django-debug-toolbar>=5.2.0",
    "django-htmx>=1.23.0",
    "django-probes>=1.7.0",
    "django-tables2>=2.7.5",
    "django-types>=0.20.0",
    "django-widget-tweaks>=1.5.0",
    "environs[django]>=14.1.1",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.9",
    "sentry-sdk>=2.29.1",
    "mozilla-django-oidc>=4.0.1",
]

[tool.ruff.lint]
select = [
  "E4", "E7", "E9",  # pycodestyle errors (default from ruff)
  "F",  # pyflakes
  "I",  # isort
  "N",  # naming conventions
  "DJ", # Django
  "T20", # print()
  "Q",  # quotes
  "A",  # overwritting builtins
  "LOG", # logging
  "PTH", # use pathlib
]

[tool.bumpver]
current_version = "2000.1"
version_pattern = "YYYY.INC1"
commit_message = "bump version -> {new_version}"
tag_message = "{new_version}"
tag_scope = "default"
pre_commit_hook = ""
post_commit_hook = ""
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
]
"ksp-naboj/__init__.py" = [
    'VERSION = "{version}"',
]
